simple refactor
using implicit def instead of implicit class still avoids allocation at
runtime
