+++ build.sbt
@@ -16,7 +16,7 @@ ThisBuild/Compile/scalacOptions ++= Seq(
   "-Xfatal-warnings",
   "-feature",
   "-deprecation",
-  // "-language:implicitConversions",
+  "-language:implicitConversions",
   // "-language:existentials",
 )
