+++ traversal/src/main/scala/io/shiftleft/overflowdb/traversal/EdgeTraversal.scala
@@ -1,0 +1,8 @@
+package io.shiftleft.overflowdb.traversal
+
+import io.shiftleft.overflowdb.{NodeRef, OdbEdge}
+
+class EdgeTraversal[A <: OdbEdge](val traversal: Traversal[A]) extends AnyVal {
+  def inV: Traversal[NodeRef[_]] =
+    traversal.map(_.inVertex.asInstanceOf[NodeRef[_]])
+}
